WIP: Server and UI are in difference processes.
The UI and Server now use communication channels from the multiprocessing
module.  This seems slightly slower than when everything is running in a
single process, but the code is much cleaner and easier to follow.

Parsing from a cache is taking about 14s, whereas it used to take about 8s.

TODO: if you Ctrl-C twice and bail during parsing, the processes don't get
cleaned up.  Need to figure out how to handle/fix this.

Signed-off-by: Bob Foerster <rfoerster@layerzero.com>